update friend list is working
